Re-org'ed ED code to solve Solaris issue.
